Move SustainabilityLabel class from domain.py to sustainability_labels.py due to circular import.
